Initial split from GEOSgcm_GridComp.
Several use cases favor the separation of this layer into a separate
repository.  The fvdycore itself is now brought in (via
manage_externals) from a fork of the GFDL repository.

Initially, there will be an orphaned branch of the fvdycore repo to
reflect the current GMAO version.  This should be reconciled in the
near future though.