Add automatic PR for develop to MAPL3 branch
MAPL will soon be working toward a version 3 that will have incompatibilities to MAPL 2. This branch is now on this repository called `release/MAPL-v3` and will allow the SI Team to make changes to this repository while not affecting `develop` directly.

What this PR does is add a GitHub Action so that any push to `develop` (aka, any PR merged into `develop`) also automatically makes a PR from `develop` into `release/MAPL-v3`.

This way, our tracking branch can "keep up" with `develop` and once MAPL 3 is ready to go, all that is needed is to merge `release/MAPL-v3` into `develop` and we are good.